Move to umbrella header project struct; remove unnamed namespace for Aho-Corasick util functions
