+++ include/algorithm.hpp
@@ -1,12 +1,6 @@
 #ifndef ALGORITHM_HPP
 #define ALGORITHM_HPP
 
-#include <vector>
-#include <string>
-#include <set>
-
-namespace algo {
-    std::vector<std::vector<size_t>> findAllStringsAhoCorasick(const std::string &text, const std::set<std::string>& patterns);
-}
+#include "AhoCorasick.hpp"
 
 #endif // ALGORITHM_HPP
